// src/generated/models/PRD_ApprovalsModel.ts
export interface Mastercard_IDValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Review_IDValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Log_IDValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface field_3Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface field_72Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface PRD_ApprovalsBase {
  ID?: number;
  Title?: string;
  PRD_Number?: number;
  Part_No?: string;
  Machine_ID?: number;
  "Mastercard_ID#Id"?: number;
  "Review_ID#Id"?: number;
  "Log_ID#Id"?: number;
  "field_3#Id"?: number;
  BPDC_Required?: boolean;
  field_5?: string;
  field_6?: number;
  field_7?: string;
  field_8?: string;
  field_9?: string;
  field_10?: number;
  field_11?: number;
  field_12?: number;
  field_13?: number;
  field_14?: number;
  field_15?: number;
  field_16?: number;
  field_17?: number;
  field_18?: number;
  field_19?: number;
  field_20?: number;
  field_21?: number;
  field_22?: number;
  field_23?: number;
  field_24?: number;
  field_25?: number;
  field_26?: number;
  field_27?: number;
  field_28?: number;
  field_29?: number;
  field_30?: number;
  field_31?: number;
  field_32?: number;
  field_33?: number;
  field_34?: number;
  field_35?: number;
  field_36?: number;
  field_37?: number;
  field_38?: number;
  field_39?: number;
  field_40?: number;
  field_41?: number;
  field_42?: number;
  field_43?: string;
  field_44?: number;
  field_45?: number;
  field_46?: number;
  field_47?: number;
  field_48?: number;
  field_49?: number;
  field_50?: number;
  field_51?: number;
  field_52?: number;
  field_53?: number;
  field_54?: number;
  field_55?: number;
  field_56?: number;
  field_57?: number;
  field_58?: number;
  field_59?: number;
  field_60?: number;
  field_61?: number;
  field_62?: number;
  field_63?: number;
  field_64?: number;
  field_65?: number;
  field_66?: number;
  field_67?: number;
  field_68?: number;
  field_69?: number;
  field_70?: number;
  field_71?: string;
  "field_72#Id"?: number;
  field_73?: string;
  field_74?: string;
  field_75?: string;
  field_76?: string;
  field_77?: number;
  field_78?: string;
  field_79?: string;
  Proposed_Shot_Weight_gr?: number;
  Propposed_Special_Instructions?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface PRD_ApprovalsRead extends PRD_ApprovalsBase {
  Mastercard_ID?: Mastercard_IDValue;
  Review_ID?: Review_IDValue;
  Log_ID?: Log_IDValue;
  field_3?: field_3Value;
  field_72?: field_72Value;
  Author?: AuthorValue;
  Editor?: EditorValue;
}